+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - 📈 now **.url** files are also supported for image preview
 - 📈 now **.exe** files are also supported for image preview
 - 📈 now **.lnk** files are also supported for image preview
+- 📈 the **ls command** now supports hyprlinks when supplied with the `--hyprlink` flag, or when the `MCAT_HYPRLINK` env var is set to either true or 1
 
 ## V0.4.5
 - 📈 **markdown viewer** now supports `<figure>` and `<figcaption>` HTML elements